# BACKEND/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel CRUD",
        "description": "Documentação da API Laravel CRUD",
        "contact": {
            "email": "seu-email@dominio.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/sellers": {
            "get": {
                "tags": [
                    "Sellers"
                ],
                "summary": "Listar todos os vendedores com suas lojas",
                "operationId": "6768a646cebc6baa1b2396f48fac0e10",
                "responses": {
                    "200": {
                        "description": "Lista de vendedores retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Sellers"
                ],
                "summary": "Criar um novo vendedor",
                "operationId": "9249d06f0fbe023b9dfb798a4a842da3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_shop",
                                    "name",
                                    "cpf"
                                ],
                                "properties": {
                                    "id_shop": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "João da Silva"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "123.456.789-00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Vendedor criado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/sellers/{seller}": {
            "get": {
                "tags": [
                    "Sellers"
                ],
                "summary": "Obter detalhes de um vendedor específico",
                "operationId": "5ad7c3f831845b3a55b6424bd8d1e740",
                "parameters": [
                    {
                        "name": "seller",
                        "in": "path",
                        "description": "ID do vendedor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vendedor retornado com sucesso"
                    },
                    "404": {
                        "description": "Vendedor não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Sellers"
                ],
                "summary": "Atualizar dados de um vendedor",
                "operationId": "4f29ed596ff352c6c60a0a72013d9846",
                "parameters": [
                    {
                        "name": "seller",
                        "in": "path",
                        "description": "ID do vendedor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "João Atualizado"
                                    },
                                    "cpf": {
                                        "type": "string",
                                        "example": "987.654.321-00"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Vendedor atualizado com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Sellers"
                ],
                "summary": "Deletar um vendedor",
                "operationId": "513eb36fc81f3789f9bbf92aed8f27fc",
                "parameters": [
                    {
                        "name": "seller",
                        "in": "path",
                        "description": "ID do vendedor",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Vendedor deletado com sucesso"
                    },
                    "404": {
                        "description": "Vendedor não encontrado"
                    }
                }
            }
        },
        "/api/shops": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Listar todas as lojas",
                "operationId": "811428430320080e8bdf49e15e1352fb",
                "responses": {
                    "200": {
                        "description": "Lista de lojas retornada com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Shops"
                ],
                "summary": "Criar uma nova loja",
                "operationId": "eefcf4c24ec6cf6cd3f50d507d3afb02",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cnpj",
                                    "cep",
                                    "address",
                                    "neighborhood",
                                    "city",
                                    "state"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Loja Central"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "12.345.678/0001-99"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "30100-000"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Rua das Flores, 100"
                                    },
                                    "neighborhood": {
                                        "type": "string",
                                        "example": "Centro"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Belo Horizonte"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "MG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Loja criada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/shops/{shop}": {
            "get": {
                "tags": [
                    "Shops"
                ],
                "summary": "Exibir uma loja específica",
                "operationId": "6465c42094d92d53b51c143ab2d37e37",
                "parameters": [
                    {
                        "name": "shop",
                        "in": "path",
                        "description": "ID da loja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Loja encontrada com sucesso"
                    },
                    "404": {
                        "description": "Loja não encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Shops"
                ],
                "summary": "Atualizar uma loja",
                "operationId": "dd26e715c342092056e0adfea5738613",
                "parameters": [
                    {
                        "name": "shop",
                        "in": "path",
                        "description": "ID da loja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "cnpj",
                                    "cep",
                                    "address",
                                    "neighborhood",
                                    "city",
                                    "state"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Loja Atualizada"
                                    },
                                    "cnpj": {
                                        "type": "string",
                                        "example": "98.765.432/0001-88"
                                    },
                                    "cep": {
                                        "type": "string",
                                        "example": "30130-000"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Av. Amazonas, 200"
                                    },
                                    "neighborhood": {
                                        "type": "string",
                                        "example": "Savassi"
                                    },
                                    "city": {
                                        "type": "string",
                                        "example": "Belo Horizonte"
                                    },
                                    "state": {
                                        "type": "string",
                                        "example": "MG"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Loja atualizada com sucesso"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Shops"
                ],
                "summary": "Excluir uma loja",
                "operationId": "db7ad626e17c631edea38504dbca8259",
                "parameters": [
                    {
                        "name": "shop",
                        "in": "path",
                        "description": "ID da loja",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Loja excluída com sucesso"
                    },
                    "404": {
                        "description": "Loja não encontrada"
                    }
                }
            }
        }
    },
...
}